requirements.txt
asgiref==3.6.0
attrs==22.2.0
colorama==0.4.6
Django==3.2.16
django-bootstrap3==22.2
django-countries==7.5
django-localflavor==3.1
django-nested-admin==4.0.2
django-reset-migrations==0.4.0
exceptiongroup==1.1.0
importlib-metadata==2.1.3
iniconfig==1.1.1
mysql-connector-python==8.0.31
packaging==22.0
pip==22.3.1
pluggy==1.0.0
protobuf==3.20.1
pytest==7.2.0
python-monkey-business==1.0.0
pytz==2022.7
setuptools==65.6.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
wheel==0.38.4
zipp==3.11.0
python-stdnum==1.18
black==22.12.0
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
typed-ast==1.5.4
